Move format specific types from formatter.ts to types.ts inside formatter folder for consistency
